+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -11,6 +11,10 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Added password input on Settings (#3035)
 - Added `date` input on custom post mutations (`createPost`, `updatePost`, `createCustomPost`, and `updateCustomPost`) (#3036)
 
+### Improvements
+
+- Removed stale polyfill interface `DateTimeInterface` (for PHP 7.2)
+
 ## 10.3.1 - 15/02/2025
 
 ### Fixed